pyproject.toml
[project]
name = "ia-com-django"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=5.1.3",
    "django-extensions>=4.1",
    "langchain-groq>=1.1.2",
    "markdown>=3.10.2",
    "python-decouple>=3.8",
    "whitenoise>=6.12.0",
]
